More generalising Shapes for transformations (fixing a vertex non-transformation bug in the process), and also beginning to work on trying to figure out what's wrong with particle pooling.

+++ src/com/stabilise/core/app/EventBus.java
@@ -169,9 +169,9 @@ public class EventBus {
 	private static final class Listener {
 		private final Consumer<Event> handler;
 		private final boolean persistent;
-		private int uses;
+		private long uses;
 		
-		public Listener(Consumer<Event> handler, boolean persistent, int uses) {
+		public Listener(Consumer<Event> handler, boolean persistent, long uses) {
 			this.handler = Objects.requireNonNull(handler);
 			this.persistent = persistent;
 			this.uses = uses;
